Working for Tauler y Fau Auction 60
Currently, it will look at all lots - you must supply the
"all lots" url - and get the links to the lots, then look
at each of the lots to get the information about the lot
and the links to images, then it gets the images.  Also,
it will save the dataframe in parquet format.

Lots to be done, but we are on the way.